[NAYB-94] refactor : ItemFixture 패키지 위치 변경
